# server/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"MoveOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"gameHistory_id": {
						"type": "number",
						"format": "double"
					},
					"isABlackPiece": {
						"type": "boolean"
					},
					"whichPiece": {
						"type": "string"
					},
					"from": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"to": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"gameHistory_id",
					"isABlackPiece",
					"whichPiece",
					"from",
					"to"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameHistoryOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"room_uuid": {
						"type": "string"
					},
					"blackPlayer_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"whitePlayer_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"winner_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"moves": {
						"items": {
							"$ref": "#/components/schemas/MoveOutputDTO"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"id",
					"room_uuid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameHistoryInputDTO": {
				"properties": {
					"room_uuid": {
						"type": "string"
					}
				},
				"required": [
					"room_uuid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameHistoryInputPatchDTO": {
				"properties": {
					"winner_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"blackPlayer_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"whitePlayer_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/gameHistories/user/{id}": {
			"get": {
				"operationId": "GetGameHistoriesByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameHistoryOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"GameHistories"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/gameHistories/room/{uuid}": {
			"get": {
				"operationId": "GetGameHistoriyByRoomId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameHistoryOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"GameHistories"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/gameHistories": {
			"post": {
				"operationId": "CreateGameHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"GameHistories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameHistoryInputDTO"
							}
						}
					}
				}
			}
		},
		"/gameHistories/{id}": {
			"delete": {
				"operationId": "DeleteGameHistory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"GameHistories"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateGameHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"GameHistories"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameHistoryInputPatchDTO"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}